fix(lens): wrap tileAura's three inputs in a useTileAura receptacle
Add a reactive id→TileAura accessor next to the pure tileAura mapper that gathers agentBucket/isUnread/staleness once; TerminalCanvas and CanvasMinimap now plug into the one socket instead of hand-wiring the same three classifiers.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
